Add a CutList data structure.
This data structure currently only contains the boards that are necessary
for purchase, as specified directly by the user. Over time, we're going to
modify this data structure to contain two separate aspects: a list of
boards that need to be purchased, in nominal form, and a listing of cuts
that need to be made to these boards in order to result in the final
configuration the user requires.

Refs #3.